+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * feat: API Platform support
 * feat(`OpenApi`): change all 'page' parameters to accept string
 * feat(ApiPlatform): add `PageNormalizer` & `PagerFactory`
+* feat(`QueryBuilderAdapter`): add type detection
 
 # 0.7.2